refactor: move Education store to hooks.ts
