+++ src/components/education/education.tsx
@@ -1,37 +1,8 @@
 import { useForm } from "react-hook-form";
-import { create } from "zustand";
 import { ChangeEvent } from "react";
 import { Summary } from "../details/summary";
 import { Details } from "../details/details";
-
-type EducationData = {
-  education: string;
-  school: string;
-  startDate: string;
-  endDate: string;
-  description: string;
-};
-
-type Action = {
-  updateEducation: (education: EducationData["education"]) => void;
-  updateSchool: (school: EducationData["school"]) => void;
-  updateStartDate: (startDate: EducationData["startDate"]) => void;
-  updateEndDate: (endDate: EducationData["endDate"]) => void;
-  updateDescription: (description: EducationData["description"]) => void;
-};
-
-const useEducationStore = create<EducationData & Action>()((set) => ({
-  education: "",
-  school: "",
-  startDate: "",
-  endDate: "",
-  description: "",
-  updateEducation: (education) => set(() => ({ education: education })),
-  updateSchool: (school) => set(() => ({ school: school })),
-  updateStartDate: (startDate) => set(() => ({ startDate: startDate })),
-  updateEndDate: (endDate) => set(() => ({ endDate: endDate })),
-  updateDescription: (description) => set(() => ({ description: description })),
-}));
+import { EducationData, useEducationStore } from "./hooks";
 
 export function Education() {
   const { register } = useForm<EducationData>();
